refactor: extract FileDialogUtilities delegates (679→307 lines)
Extract SelectedPathAutomation record and SaveDialogDiscoveryWriter
into package-private delegate classes. Public API unchanged.
Add characterization tests for escapeJson, path automation, and
discovery evidence writing.

Closes #622
